move resources into operator package
