+++ api/pkg/controller/operator-master/resources/kubermatic/ui.go
